+++ lib/src/90_aliases.dart
@@ -30,7 +30,7 @@ class Qrandom extends Xoshiro128pp {
     return Qrandom._fullSeed(r1.nextRaw32(), r2.nextRaw32(), r1.nextRaw32(), r2.nextRaw32());
   }
 
-  static Qrandom expected() => Qrandom._fullSeed(Xoshiro128pp.defaultSeedA,
+  static Qrandom seeded() => Qrandom._fullSeed(Xoshiro128pp.defaultSeedA,
       Xoshiro128pp.defaultSeedB, Xoshiro128pp.defaultSeedC, Xoshiro128pp.defaultSeedD);
 }