+++ test/node_unsuported_tests.dart
@@ -13,9 +13,9 @@ import 'package:xrandom/xrandom.dart';
 void main() {
   test('64', () {
     expect(() => Xorshift32().nextRaw64(), throwsA(isA<Unsupported64Error>()));
-    expect(() => Xorshift64.expected(), throwsA(isA<Unsupported64Error>()));
-    expect(() => Xorshift128p.expected(), throwsA(isA<Unsupported64Error>()));
-    expect(() => Xoshiro256pp.expected(), throwsA(isA<Unsupported64Error>()));
-    expect(() => Splitmix64.expected(), throwsA(isA<Unsupported64Error>()));
+    expect(() => Xorshift64.seeded(), throwsA(isA<Unsupported64Error>()));
+    expect(() => Xorshift128p.seeded(), throwsA(isA<Unsupported64Error>()));
+    expect(() => Xoshiro256pp.seeded(), throwsA(isA<Unsupported64Error>()));
+    expect(() => Splitmix64.seeded(), throwsA(isA<Unsupported64Error>()));
   });
 }