+++ test/xorshift64_test.dart
@@ -9,7 +9,7 @@ import 'package:xrandom/xrandom.dart';
 import 'helper.dart';
 
 void main() {
-  testCommonRandom(() => Xorshift64(), ()=>Xorshift64.expected());
+  testCommonRandom(() => Xorshift64(), ()=>Xorshift64.seeded());
 
   checkReferenceFiles(() => Xorshift64(1), 'a');
   checkReferenceFiles(() => Xorshift64(42), 'b');